+++ lib/ohai/plugins/azure.rb
@@ -1,3 +1,7 @@
+# Author:: Alexey Karpik <alexey.karpik@rightscale.com>
+# Author:: Peter Schroeter <peter.schroeter@rightscale.com>
+# Author:: Stas Turlo <stanislav.turlo@rightscale.com>
+# Copyright:: Copyright (c) 2010-2014 RightScale Inc
 # Copyright:: Copyright (c) 2013 Opscode, Inc.
 # License:: Apache License, Version 2.0
 #
@@ -14,22 +18,31 @@
 # limitations under the License.
 #
 
+require 'ohai/mixin/azure_metadata'
+
 Ohai.plugin(:Azure) do
   provides "azure"
 
+  include ::Ohai::Mixin::AzureMetadata
+
+
+  # Identifies the azure cloud
+  #
+  # === Return
+  # true:: If the azure cloud can be identified
+  # false:: Otherwise
+  def looks_like_azure?
+    !!hint?('azure')
+  end
+
   collect_data do
-    # The azure hints are populated by the knife plugin for Azure.
-    # The project is located at https://github.com/opscode/knife-azure
-    # Please see the lib/chef/knife/azure_server_create.rb file in that
-    # project for details
-    azure_metadata_from_hints = hint?('azure')
-    if azure_metadata_from_hints
-      Ohai::Log.debug("azure_metadata_from_hints is present.")
+    if looks_like_azure?
+      ::Ohai::Log.debug("looks_like_azure? == true")
+      metadata = fetch_azure_metadata
       azure Mash.new
-      azure_metadata_from_hints.each {|k, v| azure[k] = v }
-    else
-      Ohai::Log.debug("No hints present for azure.")
-      false
+      if metadata
+        metadata.each { |k,v| azure[k] = v }
+      end
     end
   end
 end