+++ spec/unit/plugins/azure_spec.rb
@@ -16,57 +16,28 @@
 # limitations under the License.
 #
 
-
 require File.expand_path(File.dirname(__FILE__) + '/../../spec_helper.rb')
 require 'open-uri'
 
 describe Ohai::System, "plugin azure" do
+  let (:fetched_metadata) {
+    {
+      'public_ipv4' => '1.2.3.4',
+      'private_ipv4' => '192.168.0.1',
+      'public_hostname' => 'public_hostname'
+    }
+  }
+
   before(:each) do
-    @plugin = get_plugin("azure")
+    @plugin = get_plugin('azure')
+    allow(@plugin).to receive(:hint?).with('azure').and_return({})
+    allow(@plugin).to receive(:fetch_azure_metadata).and_return(fetched_metadata)
+    @plugin.run
   end
 
-  describe "with azure cloud file" do
-    before(:each) do
-      allow(File).to receive(:exist?).with('/etc/chef/ohai/hints/azure.json').and_return(true)
-      allow(File).to receive(:read).with('/etc/chef/ohai/hints/azure.json').and_return('{"public_ip":"137.135.46.202","vm_name":"test-vm","public_fqdn":"service.cloudapp.net","public_ssh_port":"22", "public_winrm_port":"5985"}')
-      allow(File).to receive(:exist?).with('C:\chef\ohai\hints/azure.json').and_return(true)
-      allow(File).to receive(:read).with('C:\chef\ohai\hints/azure.json').and_return('{"public_ip":"137.135.46.202","vm_name":"test-vm","public_fqdn":"service.cloudapp.net","public_ssh_port":"22", "public_winrm_port":"5985"}')
-      @plugin.run
-    end
-
-    it 'should set the azure cloud attributes' do
-      expect(@plugin[:azure]).not_to be_nil
-      expect(@plugin[:azure]['public_ip']).to  eq("137.135.46.202")
-      expect(@plugin[:azure]['vm_name']).to eq("test-vm")
-      expect(@plugin[:azure]['public_fqdn']).to eq("service.cloudapp.net")
-      expect(@plugin[:azure]['public_ssh_port']).to eq("22")
-      expect(@plugin[:azure]['public_winrm_port']).to eq("5985")
-    end
-
+  # Provide only success scenario
+  it 'populate azure node with required attributes' do
+    expect(@plugin[:azure]).not_to be_nil
+    expect(@plugin[:azure]).to eq(fetched_metadata)
   end
-
-  describe "without azure cloud file" do
-    before(:each) do
-      allow(File).to receive(:exist?).with('/etc/chef/ohai/hints/azure.json').and_return(false)
-      allow(File).to receive(:exist?).with('C:\chef\ohai\hints/azure.json').and_return(false)
-    end
-
-    it 'should not behave like azure' do
-      expect(@plugin[:azure]).to be_nil
-    end
-  end
-
-  describe "with rackspace cloud file" do
-    before(:each) do
-      allow(File).to receive(:exist?).with('/etc/chef/ohai/hints/rackspace.json').and_return(true)
-      allow(File).to receive(:read).with('/etc/chef/ohai/hints/rackspace.json').and_return('')
-      allow(File).to receive(:exist?).with('C:\chef\ohai\hints/rackspace.json').and_return(true)
-      allow(File).to receive(:read).with('C:\chef\ohai\hints/rackspace.json').and_return('')
-    end
-
-    it 'should not behave like azure' do
-      expect(@plugin[:azure]).to be_nil
-    end
-  end
-
 end