+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased:
 
+* Update Azure cloud plugin to query server for data instead of reading them form hint file
+
 ## Release 8.0.0
 
 * [**sawanoboly**](https://github.com/sawanoboly)